+++ daemon/bootstrap/Cargo.toml
@@ -13,8 +13,7 @@ path = "src/main.rs"
 
 [dependencies]
 ozmux_http_server = { path = "../http_server" }
-ozmux_extension = { path = "../extension" }
-ozmux_session = { path = "../session" }
+ozmux_multiplexer = { path = "../multiplexer" }
 ozmux_terminal = { path = "../terminal" }
 tokio = { workspace = true, features = ["rt-multi-thread", "macros", "signal"] }
 tracing = { workspace = true }
