+++ daemon/http_server/Cargo.toml
@@ -8,7 +8,7 @@ authors.workspace = true
 publish.workspace = true
 
 [dependencies]
-ozmux_session = { path = "../session" }
+ozmux_multiplexer = { path = "../multiplexer" }
 ozmux_terminal = { path = "../terminal" }
 axum = { workspace = true }
 tokio = { workspace = true, features = ["rt-multi-thread", "macros"] }
