+++ daemon/multiplexer/src/window_service.rs
